[kernel][vmm] Replace uint32_t with more relevant types

+++ kernel/include/mem/vmm/vmm.h
@@ -60,34 +60,34 @@ struct dynamic_array;
 int vmm_setup();
 int vmm_setup_secondary_cpu();
 
-int vmm_allocate_ptable(uint32_t vaddr);
-int vmm_free_ptable(uint32_t vaddr, struct dynamic_array* zones);
+int vmm_allocate_ptable(uintptr_t vaddr);
+int vmm_free_ptable(uintptr_t vaddr, struct dynamic_array* zones);
 int vmm_free_pdir(pdirectory_t* pdir, struct dynamic_array* zones);
 
-int vmm_map_page(uint32_t vaddr, uint32_t paddr, uint32_t settings);
-int vmm_map_pages(uint32_t vaddr, uint32_t paddr, uint32_t n_pages, uint32_t settings);
-int vmm_unmap_page(uint32_t vaddr);
-int vmm_unmap_pages(uint32_t vaddr, uint32_t n_pages);
-int vmm_copy_page(uint32_t to_vaddr, uint32_t src_vaddr, ptable_t* src_ptable);
+int vmm_map_page(uintptr_t vaddr, uintptr_t paddr, uint32_t settings);
+int vmm_map_pages(uintptr_t vaddr, uintptr_t paddr, size_t n_pages, uint32_t settings);
+int vmm_unmap_page(uintptr_t vaddr);
+int vmm_unmap_pages(uintptr_t vaddr, size_t n_pages);
+int vmm_copy_page(uintptr_t to_vaddr, uintptr_t src_vaddr, ptable_t* src_ptable);
 
 pdirectory_t* vmm_new_user_pdir();
 pdirectory_t* vmm_new_forked_user_pdir();
-void* vmm_bring_to_kernel(uint8_t* src, uint32_t length);
-void vmm_prepare_active_pdir_for_copying_at(uint32_t dest_vaddr, uint32_t length);
-void vmm_copy_to_user(void* dest, void* src, uint32_t length);
-void vmm_copy_to_pdir(pdirectory_t* pdir, void* src, uint32_t dest_vaddr, uint32_t length);
+void* vmm_bring_to_kernel(uint8_t* src, size_t length);
+void vmm_prepare_active_pdir_for_copying_at(uintptr_t dest_vaddr, size_t length);
+void vmm_copy_to_user(void* dest, void* src, size_t length);
+void vmm_copy_to_pdir(pdirectory_t* pdir, void* src, uintptr_t dest_vaddr, size_t length);
 void vmm_zero_user_pages(pdirectory_t* pdir);
 
 pdirectory_t* vmm_get_active_pdir();
 pdirectory_t* vmm_get_kernel_pdir();
 
-int vmm_load_page(uint32_t vaddr, uint32_t settings);
-int vmm_tune_page(uint32_t vaddr, uint32_t settings);
-int vmm_tune_pages(uint32_t vaddr, uint32_t length, uint32_t settings);
-int vmm_free_page(uint32_t vaddr, page_desc_t* page, struct dynamic_array* zones);
+int vmm_load_page(uintptr_t vaddr, uint32_t settings);
+int vmm_tune_page(uintptr_t vaddr, uint32_t settings);
+int vmm_tune_pages(uintptr_t vaddr, size_t length, uint32_t settings);
+int vmm_free_page(uintptr_t vaddr, page_desc_t* page, struct dynamic_array* zones);
 
 int vmm_switch_pdir(pdirectory_t* pdir);
 
-int vmm_page_fault_handler(uint32_t info, uint32_t vaddr);
+int vmm_page_fault_handler(uint32_t info, uintptr_t vaddr);
 
 #endif // _KERNEL_MEM_VMM_VMM_H
